+++ raise_utils/interpret/__init__.py
@@ -1,2 +1,3 @@
 from raise_utils.interpret.interpret import ResultsInterpreter, DODGEInterpreter
 from raise_utils.interpret.scottknott import ScottKnott
+from raise_utils.interpret.kw import KruskalWallis