+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * Useful debug info printed when data shape mismatch occurs.
 * Add `HPO` class to add hyperopt and BOHB to `raise_utils.hyperparams`.
+* Bug fix in `FeedforwardDL`.
 
 ## 2.0.2